requirements.txt
gensim==3.8.3
matplotlib==3.1.3
nltk==3.2.5
numpy==1.16.4
pandas==0.24.2
python-weka-wrapper3==0.1.16
scikit-learn==0.23.1
scipy==1.3.0
stanza==1.1.1